requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
loguru==0.7.2
MarkupSafe==2.1.5
requests==2.31.0
typing==3.7.4.3
urllib3==2.2.1
Werkzeug==3.0.1

pip~=23.3.1
configparser~=5.3.0
attrs~=23.1.0
wheel~=0.42.0
cloudbiolinux~=0.1
cryptography~=41.0.3
Pillow~=10.0.1
docutils~=0.20.1
tornado~=6.3.3
app~=0.0.1
MyApplication~=0.1.0
config~=0.5.1
exceptiongroup~=1.1.1
chardet~=3.0.4
python-dotenv~=1.0.1